move things out of accelerated table
